+++ src/Symfony/Component/JsonStreamer/JsonStreamWriter.php
@@ -29,7 +29,10 @@ use Symfony\Component\TypeInfo\TypeResolver\TypeResolver;
 /**
  * @author Mathias Arlaud <mathias.arlaud@gmail.com>
  *
- * @implements StreamWriterInterface<array<string, mixed>>
+ * @implements StreamWriterInterface<array{
+ *     include_null_properties?: bool,
+ *     ...<string, mixed>,
+ * }>
  *
  * @experimental
  */