+++ src/Symfony/Component/JsonStreamer/CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 ---
 
  * Remove `nikic/php-parser` dependency
+ * Add `_current_object` to the context passed to value transformers during write operations
+ * Add `include_null_properties` option to encode the properties with `null` value
 
 7.3
 ---